Sets up maven build and moves to zipkin packaging conventions
