+++ .travis.yml
@@ -1,0 +1,54 @@
+# Run `travis lint` when changing this file to avoid breaking the build.
+# Default JDK is really old: 1.8.0_31; Trusty's is less old: 1.8.0_51
+# https://docs.travis-ci.com/user/ci-environment/#Virtualization-environments
+sudo: required
+dist: trusty
+
+cache:
+  directories:
+  - $HOME/.m2
+
+language: java
+
+jdk:
+  - oraclejdk8
+
+
+before_install:
+  # Parameters used during release
+  - git config user.name "$GH_USER"
+  - git config user.email "$GH_USER_EMAIL"
+  # setup https authentication credentials, used by ./mvnw release:prepare
+  - git config credential.helper "store --file=.git/credentials"
+  - echo "https://$GH_TOKEN:@github.com" > .git/credentials
+
+install:
+  # Override default travis to use the maven wrapper
+  - ./mvnw install -DskipTests=true -Dmaven.javadoc.skip=true -B -V
+
+script:
+  - ./travis/publish.sh
+
+# Don't build release tags. This avoids publish conflicts because the version commit exists both on master and the release tag.
+# See https://github.com/travis-ci/travis-ci/issues/1532
+branches:
+  except:
+    - /^[0-9]/
+
+env:
+  global:
+  # Ex. travis encrypt BINTRAY_USER=your_github_account
+  - secure: "CjMV1KSpvO4YFqc//ayF4KcHYFInpakIVEET4NTC7GbIYAqf2jCH9rzSC6PQut88Ucb10hMXYudZ7egzFAUoyiZDK9V9o8IP9XZaXSRSUJjN3jRwjU8b54i0YsPGazakES2Jq0euBbJQ1Ad4p2HbRBTwQHdSr9FGWoQblGr+tE9Lbor6QLRJudCvjhETUS8qPfnDl4BOoK8De1y1Zl25HwYKw01VNe/QHR8vgjmxKJh3wnMIwfZcTuTDlWCbviJzUP4HQUAibTNKTMkDDVMWvdwrl52QgIQ6pSgzLKcU92QyowB7ksRZqKDIfT0qUMe+RyqqQEyO+mPGivDqjqSvjNLZoS+Q7sBow3bb5fO8jTM1ncTLbgiDSDJtrPPdRxkebNQpZLlyarPGlDfAzh01g1rhsk5RxuQqBtdgiDkzdBNN3kIaFn0ubh+oy8RAzFHTzO1HNjaVxvmMoqOLQk302agFh0keSPQy6EMfWyON+RhIy5oi6rIXKwcJIbLheZAOUtmYL/Luy7upkgP3asISdaUuVQw8m+/hw4jfNwrj8e47B6K53paAzhfV1AbsMG71fLbupWbrmRng6K7UVpQGNsJtRD2jdfxNFgAAAzkwcCZ0f3oTISJZhtPkTMA6SzMIfpj5KdjV+Sq7mnLMhQmuTWWwHoxR6Xjkf32LKVMYq90="
+  # Ex. travis encrypt BINTRAY_KEY=xxx-https://bintray.com/profile/edit-xxx --add
+  - secure: "lH5VBJqy3HapYCZIyGt45TvSg/sqNTi0am+Au8Cr/VLVe7dHA1rXxLs3ujfTTGI/xRrKRIt8Nney9blvrlFnOGtvZgXWP5/vkP+sOk+beL21FizV2SuSuHSHHuKnUoVCDRoI3TxR5b4aCEMH3a8xKGEHePd+lpgf/VK0Y/LgCPyxn04nc976nEER1ZZdDiwT5EUGaWtQFRoYsO2nNmxqxnE41E8AZLGCJXEM+8jypAoVUAdSPX/HcLlooOepW5TftAVfFAXWaBcy8BUbfMeZVtKxGkqclFjionp2IaUsT++o06hqQCk8/tLcgYD2PIqjNziuPxVr62X91cVZaPetUG1w+RMmtAC6Bw2S3IZF1GQOOCLRn5bz/iEWlFui7Q7N+bvZFCPFoidz0QlCe4EqgiGTTMeYZPlD7HBVL62mO+ocpSrysNy+GBjjF2ZT78tgd1djHhgsv5VzGHCQKk5FSJ6It55qlLYuyAF9ZEmA9BRdHCrMrP4Zm1EGQsK02sEMvecGHjD/NalaCRJyKlzhzKcJrNRUcK5lyKalVl8+v/GvrINkNMmgyz0kxY2ZPmN/a6gUxqH88C5eapUWME3+uKef4mlksZtnbXXdgyDRmgdC4lZ6NuK3hme3nmmZ3l1I5uQRI09/l7DOL+wTGbS5IA3CiCHZ5wHetHAVCugtngQ="
+  # Ex. travis encrypt GH_USER_EMAIL=for_github@domain.com --add
+  - secure: "EFqIwTmERJYhi+HPwREmifKK6yFdfJkdae4JD4GCCUw7sNcmH4dWezhrbqNR87bm0GA8yEkApLbWlXNVWNuFv2Rtd+j3C+NNnwXRSRUvHIOGrWtcKry5MeCFHTHmevZD59RWp5s20Z6tuziPSBU0qWD6sNOWPW60sJcuDnihy5eLxco2Rwk/Og6qlIIxUaQHyQ+UT1yRyvx1eVr58NnptFw1S06nx3qfZrAxLpWHeL6f1UCMkgexiXTgMYZApd8Y6KYoAGJjvKuMENPgS3q6/UBAZQSdLGS84Dy/TQSoPdxrlGiCDXbVuTltO45tz9Ux0AE42fGvn24zAhCvUAbQY51hwM5NrzMdVmsF5U3dNBSPAdHiHxnIZZkzT8z0Kyy7vNtLroVOFK1Z+XoiL976tIWIFoS48yIg9PFi9TOzJb2GUd5BCiuvTHkpYj0QU0JRvJHkGHtpMf5cmxewViDscZ1sozDznVQclMLF/TEQvxKJ6SAwZGteKIb39Q6/96GNBrE1k2VBle45VeqhNRTgkEdsr+jte+zGvBp5ffqBEaNhfyOt2UC0d4E6LK1n2IynRxWO4weindnab6N9cNKtxN0tNioWqlHbq4AkyLE0USoqZJHvIKL0B9WFBhNUU0pJIDEN8ojdCWyW86SU1x/FqU39/OGNxsfSAwT8cs1542M="
+  # Ex. travis encrypt GH_USER=your_github_account --add
+  - secure: "B3vw6OHio3iJy7pA08vJq29uhlKb9mVjg8QDs1K/6EeWdfdBfocPXdAF02eamHztrDquKM03ZAulLzn9EjFdU/Om6h9gx/GDj7DQ5CE7EVIjsyu6kNW8NefZJ3hFAxUul8xnCyPZRWQ8bwDo7Mz/SyhTGClwicaH/csAOB8BsHA8sKCY4xm/1GUyyW1/X5DTsRTlVL2Zr3jwSGRWDrLZSe0yBhnZoMlaYo6tDPdZKQ/FwBH+953PoVEqBiYq9Iqy+BqA/b7+jNgIJloloiEIP+rxagcxNlCi2O55vTFBa1xsbsnT3y7+gRwl4ZdXGry6DfJ/Zqf/9XMBnUbIz5/sUT5i9bzxWlNtGQVftbEpYlFSziHVJiMigOX18dTW++OMXD6jjhtoPo/Nk/4uq73WFYuZHaTIp3EvMNJTko6MFETKRF0Et6aG24sIFbZLcJBHfzfe6FJrFTxzNedxV7hezoIpOYpQw/7lKk89DPui2tfI4bdPY6wDKnsDE5+qzU3mKgqhjRq/X/caJJbAtS21eqt5I7FuP8gbZdPjo2og3NVctWcqzSpvReIc13VeZuVMWVmJ6yDFYtvOvXHEj6wG7wKX3F9MMkegw4ovPVwYnEynLs1MoStC0adZO54FC3SIg5JGgU+TZomERWq+m8EUEzzEgW/IE/cfQ01gfChQDdU="
+  # Ex. travis encrypt GH_TOKEN=XXX-https://github.com/settings/tokens-XXX --add
+  - secure: "fXcqOTKTJAYdZzIkKXMd7aRpdYAnOJLWB/RR18cM0N5TGkKXtrVf/O1gbwQqEYp6CmQMmd1TE36/4KBCVAdb+6dkRs60usGty2iVhFjVXtpms/MV92PFoLri4S4+1BaF7QY8+6+joBeEDBik3B9rIr0jjZ4TZD8EjtQ4sGX22Q5wqB6o2S/a5GU20qpK8/+79NIchGwS0dzCHWkFkU416/BsxmV0nLYERGmBzg+KgeBZLPRSIr85Zkzx/PupnIz52qRJUTXc3r7Nwx9lZwQ95fMBjG/ouJDoK74tFJSNL2IVhR3bLkNGMvX6axj+4Zy+FJrN7ghoMZ6OsWsG9vxPQTtj5EKSaC7EIPDj80+rbBKa9jA9WqfL82sabsJoZTiRVk7gsQlWX6v6FuohkO/kb8viLBwyOjLnmH1f9kpL3BV8D4usY74QMAPb0lYj12O2mZktXSMf72tIk+9AkRviXXJiuTx3ed1fWjyWBCrDHt8PcSoek0+DVopPpzRTrlGzc9/qcgWJ0UvyPWLn06wMS+U9LKwOnjwuDBTH4pMPKGCBGMTLANrLyTFiRG4A+yFlwGl2+ajnydJn6Q3gPWpCrpm47MbebHpc/ed7zaFTvbMljGcEsLm9jhaXbHG2gK6OMOO+5zFLtuL2nZ02LznTu8D4Q2zi6hAE1XjFC0oim6w="
+  # Ex. travis encrypt SONATYPE_USER=your_sonatype_account
+  - secure: "r/PfsDfAiMDWpxMNYSmYa8V04zou5fAcnfvgZEcm8e23icjPpOubpQtRBsItwJAbdzHQEr9YXD7DcklPtWALRktaH6lJcy6W1/uxEVNYTqnx8s8vYNNAYPz88Xp/cKXTBoj5c/K5FwBMp2RtHzjn5kh9e1Gw2Rdz8MUMH68iBsyePntsysuXDyOg4wEpEtiCxIqUaoEGGZaCG3MZeQYmyyp3eRkqkdMHMxS5nBDq3oG8rLa+iu634do89tjDXjkW8sQbFDZifKuBsdgIdBxAtbOY4vsH1bCOlxyjYG7+EDkCZqWCcV2pXfdTovD+i//TMZycWk8hp76Cn9vMxW5Vh2Ggb1lAfGzVGDdIeide5k/gB2i3rqV/izqMfIlLKRGkuUyNJSFh6YtCNZDQ9nC1pK2l0A0QVpgN2k/rpBQsLjj0awkH5BZwRwM4mWkwJbr+klVG2L4N8MHcR0j+ROoRVl1g/dfpuEkr2/JeplPTE6cDxa2a/yv6XWdZslLA0AueX6jBVZM1dxZwxgQEeqvQODQ8+C7BtJCfaa3R2bnbqQ6smZM+WrQjJ0Bf+/fWsM2YTL5fUbvMKn8NQOv2mMqR3nCzW26uqzBoJgoHyfcwgaCtOsgelOBzKUOhtda2ifHuoDnn4P36umq+dHe/T3I+AFkIGVb+tK3j4nUcMJW+YVk="
+  # Ex. travis encrypt SONATYPE_PASSWORD=your_sonatype_password
+  - secure: "B9iXeqXlL5AIMQ9jATVaK6J6Uy1oSOR9Jl42dZsEiK6pc6ZHGpf0zAoStApUPp5/GjtSXzPObnYEu+R7FVOHlOhcPwFdY71SyD4haSCnOFTIBPJbkCunIAA94u6jO1E/AORiETCF31zsl8d7neXUFougCzpvpVvC9PrSZceiHlCfI1+ggVWJKrJ2/C9dzfuM4pdk2jUB/29uq8OWYkWbPfW9PFbBHT91Pt/I1gKKb9jBzbD6ROeha/2CdFsBlk2gLqMoWdrDcD/nwy0GiuRzR2LFCjmclxf/euSkLcGYDRzWnzam7t4U9a2ShCFnp/EuxT4hxF2vEkBs+GS4PzFkQK8Ow4aqrkHiNMYWRluYSxBQ8VcPZnbJalepqYHSmRakE8OPakELIojCRT2rg2nK4Xgk1m1u4qFir0atjuP878Z7JQotN82P5OvuDkWb5D3APDDHq1St+9eVXPrtQe4TZr2cp1YJ9HmuOpyzmblYWKjn61Rjd4+0EamIXylxgo6yMeIqqUJkLOAJpE1g8YT8t6uENpCgfKkNx4okSIz7cL1jlYrHX3Ac6K15UZjbkAtk0ol+HnuoKdcPSQaA9JkEf10+/1FKLPLW31R+N412NaGVxut+sY0/jnTYOClfxQxUQqceXiIF3fLYtrbTY64uIDegVDzEQANugQkrmq3WgDU="
+
