--- finagle-zipkin-kafka/src/main/resources/BUILD
@@ -1,3 +1,0 @@
-resources(name='resources',
-  sources=rglobs('*', exclude=[rglobs('BUILD*', '*.pyc')])
-)
