--- finagle-zipkin-kafka/src/main/resources/META-INF/services/com.twitter.finagle.tracing.Tracer
@@ -1,1 +1,0 @@
-com.twitter.finagle.zipkin.kafka.KafkaZipkinTracer
