--- finagle-zipkin-kafka/src/test/scala/BUILD
@@ -1,13 +1,0 @@
-junit_tests(name='scala',
-  dependencies=[
-    '3rdparty/jvm/org/mockito:mockito-all',
-    '3rdparty/jvm/org/scalatest',
-    '3rdparty/jvm/junit',
-    'util/util-core',
-    'util/util-codec',
-    'finagle/finagle-core',
-    'finagle/finagle-zipkin'
-  ],
-  fatal_warnings=True,
-  sources=rglobs('*.scala')
-)
