+++ kafka/src/main/java/zipkin/finagle/InternalJavaDocDummy.java
@@ -1,0 +1,10 @@
+package zipkin.finagle;
+
+/**
+ * This file is here to allow maven central deploys and avoid the complexity of trying to convert
+ * scala to javadoc.
+ *
+ * See https://issues.sonatype.org/browse/OSSRH-2932
+ */
+public class InternalJavaDocDummy {
+}
