+++ kafka/src/main/resources/META-INF/services/com.twitter.finagle.tracing.Tracer
@@ -1,0 +1,1 @@
+zipkin.finagle.kafka.KafkaZipkinTracer
