+++ kafka/src/test/scala/zipkin/finagle/kafka/KafkaRawZipkinTracerTest.scala
@@ -1,0 +1,134 @@
+package zipkin.finagle.kafka
+
+import java.util
+import java.util.concurrent.{Future => JFuture}
+
+import com.twitter.finagle.stats.InMemoryStatsReceiver
+import com.twitter.finagle.tracing.{Flags, SpanId, TraceId}
+import com.twitter.finagle.zipkin.core.{BinaryAnnotation, Endpoint, Span, ZipkinAnnotation}
+import com.twitter.util.{Base64StringEncoder, Time}
+import org.apache.kafka.clients.producer.{Callback, RecordMetadata}
+import org.apache.kafka.common.{Metric, MetricName, PartitionInfo}
+import org.junit.runner.RunWith
+import org.scalatest.FunSuite
+import org.scalatest.junit.JUnitRunner
+import zipkin.finagle.kafka.KafkaRawZipkinTracer.{ZipkinProducer, ZipkinProducerRecord}
+
+@RunWith(classOf[JUnitRunner])
+class KafkaRawZipkinTracerTest extends FunSuite {
+
+  val traceId = TraceId(Some(SpanId(123)), Some(SpanId(123)), SpanId(123), None, Flags().setDebug)
+
+  test("send properly formulated messages to kafka") {
+    val topic = "zipkin-test"
+    val producer = new TestProducer
+    val statsReceiver = new InMemoryStatsReceiver
+    val tracer = new KafkaRawZipkinTracer(producer, topic, statsReceiver)
+
+    val localEndpoint = Endpoint(2323, 23)
+    val remoteEndpoint = Endpoint(333, 22)
+
+    val annotations = Seq(
+      ZipkinAnnotation(Time.fromSeconds(123), "cs", localEndpoint),
+      ZipkinAnnotation(Time.fromSeconds(126), "cr", localEndpoint),
+      ZipkinAnnotation(Time.fromSeconds(123), "ss", remoteEndpoint),
+      ZipkinAnnotation(Time.fromSeconds(124), "sr", remoteEndpoint),
+      ZipkinAnnotation(Time.fromSeconds(123), "llamas", localEndpoint)
+    )
+
+    val span = Span(
+      traceId = traceId,
+      annotations = annotations,
+      _serviceName = Some("hickupquail"),
+      _name=Some("foo"),
+      bAnnotations = Seq.empty[BinaryAnnotation],
+      endpoint = localEndpoint)
+
+    val expected = Base64StringEncoder.decode(
+      """DAAAAAEKAAEAAAAAAAAAewsAAwAAAANmb28KAAQAAAAAAAAAewoABQAAAAAAAAB7DwAGDAA
+        |AAAUKAAEAAAAAB1TUwAsAAgAAAAJjcwwAAwgAAQAACRMGAAIAFwsAAwAAAAtoaWNrdXBxdW
+        |FpbAAACgABAAAAAAeCm4ALAAIAAAACY3IMAAMIAAEAAAkTBgACABcLAAMAAAALaGlja3Vwc
+        |XVhaWwAAAoAAQAAAAAHVNTACwACAAAAAnNzDAADCAABAAABTQYAAgAWCwADAAAAC2hpY2t1
+        |cHF1YWlsAAAKAAEAAAAAB2QXAAsAAgAAAAJzcgwAAwgAAQAAAU0GAAIAFgsAAwAAAAtoaWN
+        |rdXBxdWFpbAAACgABAAAAAAdU1MALAAIAAAAGbGxhbWFzDAADCAABAAAJEwYAAgAXCwADAA
+        |AAC2hpY2t1cHF1YWlsAAACAAkBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        |AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        |AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        |AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=""".stripMargin)
+
+    tracer.sendSpans(Seq(span))
+
+    assert(producer.messages.length == 1)
+    val msg = producer.messages.head
+    assert(msg.topic == topic)
+    assert(util.Arrays.equals(msg.value, expected))
+    assert(statsReceiver.counters == Map(List("log_span", "ok") -> 1))
+  }
+
+  test("increment error counter on synchronous failure") {
+    val topic = "zipkin-test"
+    val statsReceiver = new InMemoryStatsReceiver
+    val producer = new TestProducer {
+      override def send(record: ZipkinProducerRecord, callback: Callback): JFuture[RecordMetadata] = {
+        callback.onCompletion(null, new TestProducerException())
+        null
+      }
+    }
+    val tracer = new KafkaRawZipkinTracer(producer, topic, statsReceiver)
+
+    val span = Span(
+      traceId = traceId,
+      annotations = Seq.empty,
+      _serviceName = Some("hickupquail"),
+      _name=Some("foo"),
+      bAnnotations = Seq.empty[BinaryAnnotation],
+      endpoint = Endpoint(2323, 23))
+
+    tracer.sendSpans(Seq(span))
+
+    assert(statsReceiver.counters ==
+      Map(List("log_span", "error",
+        classOf[TestProducerException].getName()) -> 1))
+  }
+
+  test("increment error counter on asynchronous failure") {
+    val topic = "zipkin-test"
+    val statsReceiver = new InMemoryStatsReceiver
+    val producer = new TestProducer {
+      override def send(record: ZipkinProducerRecord, callback: Callback): JFuture[RecordMetadata] = {
+        throw new TestProducerException()
+      }
+    }
+    val tracer = new KafkaRawZipkinTracer(producer, topic, statsReceiver)
+
+    val span = Span(
+      traceId = traceId,
+      annotations = Seq.empty,
+      _serviceName = Some("hickupquail"),
+      _name=Some("foo"),
+      bAnnotations = Seq.empty[BinaryAnnotation],
+      endpoint = Endpoint(2323, 23))
+
+    tracer.sendSpans(Seq(span))
+
+    assert(statsReceiver.counters ==
+      Map(List("log_span", "error",
+        classOf[TestProducerException].getName()) -> 1))
+  }
+}
+
+final class TestProducerException extends Exception
+
+private[this] class TestProducer extends ZipkinProducer {
+  var messages: Seq[ZipkinProducerRecord] = Seq.empty
+  override def partitionsFor(topic: String): util.List[PartitionInfo] = ???
+  override def metrics(): util.Map[MetricName, _ <: Metric] = ???
+  override def send(record: ZipkinProducerRecord): JFuture[RecordMetadata] = ???
+  override def send(record: ZipkinProducerRecord, callback: Callback): JFuture[RecordMetadata] = {
+    messages ++= Seq(record)
+    callback.onCompletion(null, null)
+    null
+  }
+  override def close(): Unit = ???
+}
+
